% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{bing_traffic}
\alias{bing_traffic}
\title{Get traffic raster from Bing Maps}
\usage{
bing_traffic(latitude, longitude, height, width, zoom, bing_key)
}
\arguments{
\item{latitude}{Latitude of center of area}

\item{longitude}{Longitude of center of area}

\item{height}{Height}

\item{width}{Width}

\item{zoom}{Zoom; integer from 0 to 20. For more information, see \href{https://wiki.openstreetmap.org/wiki/Zoom_levels}{here}}

\item{bing_key}{Bing API key}
}
\value{
Returns a georeferenced raster file. The file can contain the following values: 1 = no traffic; 2 = light traffic; 3 = moderate traffic; 4 = heavy traffic.
}
\description{
This function returns a raster of traffic levels from Bing Maps.
}
