% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{det_google_pixel_dist_deg}
\alias{det_google_pixel_dist_deg}
\title{Determine pixel distance in decimal degrees}
\usage{
det_google_pixel_dist_deg(zoom)
}
\arguments{
\item{zoom}{Zoom level}
}
\value{
Returns the pixel height/width in decimal degrees (integer)
}
\description{
Based on the zoom level, determine the height/width of each pixel in decimal degrees
}
