% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt_estimate_webshot_delay.R
\name{gt_estimate_webshot_delay}
\alias{gt_estimate_webshot_delay}
\title{If not specified, estimate webshot_delay using height and width}
\usage{
gt_estimate_webshot_delay(height, width, webshot_delay)
}
\arguments{
\item{height}{Height}

\item{width}{Width}

\item{webshot_delay}{Webshot Delay}
}
\value{
webshot_delay (in seconds).
}
\description{
If not specified, estimate webshot_delay using height and width
}
