% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt_html_to_raster.R
\name{gt_html_to_raster}
\alias{gt_html_to_raster}
\title{Converts Google HTML file to Raster}
\usage{
gt_html_to_raster(
  filename,
  location = NULL,
  height = NULL,
  width = NULL,
  zoom = NULL,
  webshot_delay = 10,
  save_png = F
)
}
\arguments{
\item{height}{Height}

\item{width}{Width}

\item{zoom}{Zoom level
\subsection{Other parameters}{
}}

\item{webshot_delay}{How long to wait for .html file to load. Larger .html files will require more time to fully load.}

\item{save_png}{The function creates a .png file as an intermediate step. Specify whether the .png file should be kept (default: \code{FALSE})}

\item{polygon}{\code{SpatialPolygonsDataframe} the defines region to be queried.
\subsection{If \code{save_params} is set to \code{FALSE} in `` or \code{gt_make_htmls_from_grid}, then the following must be specified}{
}}
}
\value{
Returns a georeferenced raster file. The file can contain the following values: 1 = no traffic; 2 = light traffic; 3 = moderate traffic; 4 = heavy traffic.
}
\description{
Converts a Google HTML file into a spatially referenced raster file.
}
