% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{gt_htmls_to_raster}
\alias{gt_htmls_to_raster}
\title{Converts multiple Google HTML files into a raster}
\usage{
gt_htmls_to_raster(
  html_files,
  webshot_delay,
  grid_param_df = NULL,
  save_png = F,
  print_progress = T
)
}
\arguments{
\item{html_files}{Vector of html_files}

\item{webshot_delay}{How long to wait for .html file to load. Larger .html files will require more time to fully load.
\subsection{If \code{save_params} is set to \code{FALSE} in \code{gt_make_htmls_from_grid}, then the following must be specified}{
}}

\item{grid_param_df}{Grid parameter dataframed defined by \code{gt_make_point_grid()}
\subsection{Other parameters}{
}}

\item{save_png}{The function creates a .png file as an intermediate step. Specify whether the .png file should be kept (default: \code{FALSE})}

\item{print_progress}{Whether to print progress to show which file the function is processing (default: \code{TRUE})}
}
\value{
Returns a georeferenced raster file. The file can contain the following values: 1 = no traffic; 2 = light traffic; 3 = moderate traffic; 4 = heavy traffic.
}
\description{
Converts a multiple Google HTML files into a single spatially referenced raster file.
}
