% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt_load_png_as_traffic_raster.R
\name{gt_load_png_as_traffic_raster}
\alias{gt_load_png_as_traffic_raster}
\title{Converts png to raster}
\usage{
gt_load_png_as_traffic_raster(
  filename,
  latitude,
  longitude,
  height,
  width,
  zoom
)
}
\arguments{
\item{img}{PNG object from \code{readPNG}}
}
\value{
Returns a raster
}
\description{
Converts PNG to raster and translates color values to traffic values
}
