% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt_make_extent.R
\name{gt_make_extent}
\alias{gt_make_extent}
\title{Determine the spatial extent of a Google traffic tile}
\usage{
gt_make_extent(latitude, longitude, height, width, zoom)
}
\arguments{
\item{latitude}{Latitude}

\item{longitude}{Longitude}

\item{height}{Height}

\item{width}{Width}

\item{zoom}{Zoom level}
}
\value{
Returns an extent object in WGS84
}
\description{
Based on the location, height, width, and zoom, determines the spatial extent of the Google traffic tile
}
