% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{gt_make_htmls_from_grid}
\alias{gt_make_htmls_from_grid}
\title{Make multiple traffic html from Google based on grid of points}
\usage{
gt_make_htmls_from_grid(
  grid_param_df,
  filename_suffix,
  out_dir,
  google_key,
  save_params = F
)
}
\arguments{
\item{google_key}{Google API key}

\item{save_params}{Save an .Rds file that contains the parameters (location, height, width and zoom). This file can then be used by the \code{gt_html_to_raster()} function.}

\item{location}{Vector of latitude and longitude}

\item{height}{Height}

\item{width}{Width}

\item{zoom}{Zoom; integer from 0 to 20. For more information, see \href{https://wiki.openstreetmap.org/wiki/Zoom_levels}{here}}

\item{filename}{Path and filename to save file}
}
\value{
Returns an html file of Google traffic
}
\description{
This function returns multiple html of traffic from Google based on a grid defined by \code{gt_make_point_grid}. The \code{gt_htmls_to_raster()} can then be used to these html files into a georeferenced raster file.
}
