% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt_make_point_grid.R
\name{gt_make_point_grid}
\alias{gt_make_point_grid}
\title{Creates grid of points to query Google Traffic}
\usage{
gt_make_point_grid(polygon, height, width, zoom, reduce_hw = 10)
}
\arguments{
\item{polygon}{\code{SpatialPolygonsDataframe} the defines region to be queried.}

\item{height}{Height}

\item{width}{Width}

\item{zoom}{Zoom level}

\item{reduce_hw}{Number of pixels to reduce height/width by. The tiles produced by the function may not exactly overlap. Reducing the height and width ensures overlap to eventually remove any blank space.}
}
\value{
Returns a dataframe with the locations to query and parameters.
}
\description{
Querying too large of a location may be unfeasible; consequently, it may be necessary to query multiple smaller locations. Based on the location to be queried and the height, width and zoom parameters, determines the points that should be queried.
}
