% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt_make_raster.R
\name{gt_make_raster}
\alias{gt_make_raster}
\title{Make Google Traffic Raster}
\usage{
gt_make_raster(location, height, width, zoom, google_key, webshot_delay)
}
\arguments{
\item{location}{Vector of latitude and longitude}

\item{height}{Height}

\item{width}{Width}

\item{zoom}{Zoom level; integer from 0 to 20. For more information, see \href{https://wiki.openstreetmap.org/wiki/Zoom_levels}{here}}

\item{google_key}{Google API key}

\item{webshot_delay}{How long to wait for google traffic layer to render. Larger height/widths require longer delay times.}
}
\value{
Returns a georeferenced raster file. The file can contain the following values: 1 = no traffic; 2 = light traffic; 3 = moderate traffic; 4 = heavy traffic.
}
\description{
Make a raster from Google traffic data, where each pixel has one of four values
indicating traffic volume (no traffic, light, moderate, and heavy).
}
