% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt_html_to_raster.R
\name{gt_html_to_raster}
\alias{gt_html_to_raster}
\title{Converts Google HTML file to Raster}
\usage{
gt_html_to_raster(
  filename,
  location,
  height,
  width,
  zoom,
  webshot_delay = NULL,
  print_progress = T
)
}
\arguments{
\item{filename}{HTML filename to convert into raster}

\item{location}{Vector of latitude and longitude}

\item{height}{Height (in pixels; pixel length depends on zoom)}

\item{width}{Width (in pixels; pixel length depends on zoom)}

\item{zoom}{Zoom level}

\item{webshot_delay}{How long to wait for .html file to load. Larger .html files (large height/widths) will require more time to fully load. If \code{NULL}, the following delay time (in seconds) is used: \code{delay = max(height,width)/200}.}

\item{print_progress}{Whether to print function progress}
}
\value{
Returns a georeferenced raster file. The file can contain the following values: 1 = no traffic; 2 = light traffic; 3 = moderate traffic; 4 = heavy traffic.
}
\description{
Converts a Google HTML file into a spatially referenced raster file.
}
