% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt_load_png_as_traffic_raster.R
\name{gt_load_png_as_traffic_raster}
\alias{gt_load_png_as_traffic_raster}
\title{Converts png to raster}
\usage{
gt_load_png_as_traffic_raster(
  filename,
  latitude,
  longitude,
  height,
  width,
  zoom
)
}
\arguments{
\item{filename}{Filename/path of png file}

\item{latitude}{Latitude used to create png file using \code{gt_make_png()}}

\item{longitude}{Longitude used to create png file using \code{gt_make_png()}}

\item{height}{Height (in pixels; pixel length depends on zoom) used to create png file using \code{gt_make_png()}}

\item{width}{Width (in pixels; pixel length depends on zoom) used to create png file using \code{gt_make_png()}}

\item{zoom}{Zoom used to create png file using \code{gt_make_png()}}
}
\value{
Returns a raster where each pixel represents traffic level (1 = no traffic, 2 = medium traffic, 3 = traffic delays, 4 = heavy traffic)
}
\description{
Converts PNG to raster and translates color values to traffic values
}
