% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt_make_html.R
\name{gt_make_html}
\alias{gt_make_html}
\title{Make traffic html from Google}
\usage{
gt_make_html(location, height, width, zoom, filename, google_key)
}
\arguments{
\item{location}{Vector of latitude and longitude}

\item{height}{Height (in pixels; pixel length depends on zoom)}

\item{width}{Width (in pixels; pixel length depends on zoom)}

\item{zoom}{Zoom; integer from 0 to 20. For more information, see \href{https://wiki.openstreetmap.org/wiki/Zoom_levels}{here}}

\item{filename}{Path and filename to save file}

\item{google_key}{Google API key}
}
\value{
Returns an html file of Google traffic
}
\description{
This function returns an html of traffic from Google. The \code{gt_html_to_raster()} can
then be used to convert this html into a georeferenced raster file.
}
