% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt_make_png.R
\name{gt_make_png}
\alias{gt_make_png}
\title{Make Google Traffic PNG}
\usage{
gt_make_png(
  location,
  height,
  width,
  zoom,
  webshot_delay,
  google_key,
  out_filename
)
}
\arguments{
\item{location}{Vector of latitude and longitude}

\item{height}{Height}

\item{width}{Width}

\item{zoom}{Zoom level; integer from 0 to 20. For more information, see \href{https://wiki.openstreetmap.org/wiki/Zoom_levels}{here}}

\item{webshot_delay}{How long to wait for google traffic layer to render. Larger height/widths require longer delay times.}

\item{google_key}{Google API key}

\item{out_filename}{Filename/path of png file to make}
}
\value{
Returns a georeferenced raster file. The file can contain the following values: 1 = no traffic; 2 = light traffic; 3 = moderate traffic; 4 = heavy traffic.
}
\description{
Make a png file of Google traffic data. The \code{gt_load_png_as_traffic_raster()} function can then
be used to convert the png into a traffic raster
}
