% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt_make_raster_from_grid.R
\name{gt_make_raster_from_grid}
\alias{gt_make_raster_from_grid}
\title{Make Google Traffic Raster Based on Grid of Coordinates}
\usage{
gt_make_raster_from_grid(
  grid_param_df,
  webshot_delay,
  google_key,
  print_progress = T
)
}
\arguments{
\item{grid_param_df}{Grid parameter dataframe produced from \link{gt_make_grid}}

\item{webshot_delay}{How long to wait for google traffic layer to render. Larger height/widths require longer delay times.}

\item{google_key}{Google API key}

\item{print_progress}{Show progress for which tile has been processed.}
}
\value{
Returns a georeferenced raster file. The file can contain the following values: 1 = no traffic; 2 = light traffic; 3 = moderate traffic; 4 = heavy traffic.
}
\description{
Make a raster from Google traffic data, where each pixel has one of four values
indicating traffic volume (no traffic, light, moderate, and heavy).
}
