% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt_make_raster_from_grid.R
\name{gt_make_raster_from_grid}
\alias{gt_make_raster_from_grid}
\title{Make Google Traffic Raster Based on Grid of Coordinates}
\usage{
gt_make_raster_from_grid(
  grid_param_df,
  google_key,
  webshot_delay = NULL,
  return_list_of_tiles = F,
  print_progress = T
)
}
\arguments{
\item{grid_param_df}{Grid parameter dataframe produced from \link[gt_make_grid()]{\code{gt_make_grid()}}}

\item{google_key}{Google API key}

\item{webshot_delay}{How long to wait for google traffic layer to render. Larger height/widths require longer delay times. If \code{NULL}, the following delay time (in seconds) is used: \code{delay = max(height,width)/200}.}

\item{return_list_of_tiles}{Instead of merging traffic tiles together into one large tile, return a list of tiles}

\item{print_progress}{Whether to print function progress}
}
\value{
Returns a georeferenced raster file. The file can contain the following values: 1 = no traffic; 2 = light traffic; 3 = moderate traffic; 4 = heavy traffic.
}
\description{
Make a raster from Google traffic data, where each pixel has one of four values
indicating traffic volume (no traffic, light, moderate, and heavy).
}
\examples{
\dontrun{
## Grab polygon of Manhattan
us_sp <- raster::getData('GADM', country='USA', level=2)
ny_sp <- us_sp[us_sp$NAME_2 \%in\% "New York",]

## Make Grid
grid_df <- gt_make_grid(polygon = ny_sp,
                       height   = 2000,
                       width    = 2000,
                       zoom     = 16)

## Make raster from grid                        
r <- gt_make_raster_from_grid(grid_param_df = grid_clean_df,
                              google_key    = "GOOGLE-KEY-HERE")
}

}
