% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt_make_raster_from_polygon.R
\name{gt_make_raster_from_polygon}
\alias{gt_make_raster_from_polygon}
\title{Make Google Traffic Raster Based on Polygon}
\usage{
gt_make_raster_from_polygon(
  polygon,
  height,
  width,
  zoom,
  webshot_delay,
  google_key,
  reduce_hw = 10,
  return_list_of_tiles = F,
  crop_to_polygon = T,
  print_progress = T
)
}
\arguments{
\item{polygon}{Polygon (\code{sf} object or \code{SpatialPolygonsDataframe}) in WGS84 CRS}

\item{height}{Height}

\item{width}{Width}

\item{zoom}{Zoom level; integer from 0 to 20. For more information, see \href{https://wiki.openstreetmap.org/wiki/Zoom_levels}{here}}

\item{webshot_delay}{How long to wait for google traffic layer to render. Larger height/widths require longer delay times.}

\item{google_key}{Google API key}

\item{reduce_hw}{Number of pixels to reduce height/width by. Doing so creates some overlap between tiles to ensure there is not blank space between tiles (default = 10 pixels).}

\item{print_progress}{Show progress for which tile has been processed.}
}
\value{
Returns a georeferenced raster file. The file can contain the following values: 1 = no traffic; 2 = light traffic; 3 = moderate traffic; 4 = heavy traffic.
}
\description{
Make a raster from Google traffic data, where each pixel has one of four values
indicating traffic volume (no traffic, light, moderate, and heavy).
}
